+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Disable dynamic sampling and metrics extraction in Managed relays. For users with a dynamic sampling rate < 1, this increases the outgoing request volume of the Managed relay. ([#6413](https://github.com/getsentry/relay/pull/6413), [#6425](https://github.com/getsentry/relay/pull/6425))
 - Copy the INP interaction type to the INP web vital metric. ([#6420](https://github.com/getsentry/relay/pull/6420))
 - Extract session updates from Switch dying messages. ([#6363](https://github.com/getsentry/relay/pull/6363))
+- Add `spool.envelopes.max_unspool_envelopes_per_second` to limit unspooling rate. Defaults to 1000 envelopes per second per Relay instance. ([#6433](https://github.com/getsentry/relay/pull/6433))
 
 **Bug Fixes**: